Add a "ignore" label to skip a service
